requirements.txt
backports.tarfile==1.1.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.15.4
Flask==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
Jinja2==3.1.4
keyring==25.1.0
MarkupSafe==2.1.5
more-itertools==10.2.0
paho-mqtt==1.6.1
platformdirs==4.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
requests==2.32.3
six==1.16.0
urllib3==2.2.2
virtualenv==20.26.3
Werkzeug==3.0.3
zipp==3.18.1
